feat: hooks로직 변경 , 타입 분리 컴퍼넌트 분리 및 생성 -
